# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Andy Homecare Connect API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  employer: 'employer',
  housekeeper: 'housekeeper',
} as const;

export interface User {
  id: number;
  fullName: string;
  role: UserRole;
  county: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** Monthly salary expectation in KES */
  salaryExpectation: number;
  /**
     * Comma-separated skills (for housekeepers)
     * @nullable
     */
  skills?: string | null;
  /**
     * Years of experience description
     * @nullable
     */
  experience?: string | null;
  /** @nullable */
  languages?: string | null;
  /**
     * Immediate, 2 weeks notice, etc.
     * @nullable
     */
  availability?: string | null;
  paymentVerified: boolean;
  registeredAt: string;
}

export type UserInputRole = typeof UserInputRole[keyof typeof UserInputRole];


export const UserInputRole = {
  employer: 'employer',
  housekeeper: 'housekeeper',
} as const;

export interface UserInput {
  /** @minLength 2 */
  fullName: string;
  role: UserInputRole;
  /** @minLength 2 */
  county: string;
  /** @minLength 10 */
  phone: string;
  email?: string;
  bio?: string;
  photoUrl?: string;
  /** @minimum 1 */
  salaryExpectation: number;
  skills?: string;
  experience?: string;
  languages?: string;
  availability?: string;
}

export interface UserUpdate {
  /** @minLength 2 */
  fullName?: string;
  county?: string;
  phone?: string;
  email?: string;
  bio?: string;
  photoUrl?: string;
  /** @minimum 1 */
  salaryExpectation?: number;
  skills?: string;
  experience?: string;
  languages?: string;
  availability?: string;
}

export type PaymentStatus = typeof PaymentStatus[keyof typeof PaymentStatus];


export const PaymentStatus = {
  pending: 'pending',
  verified: 'verified',
  rejected: 'rejected',
} as const;

export interface Payment {
  id: number;
  userId: number;
  /** Mpesa transaction code */
  mpesaCode: string;
  /** Amount paid in KES */
  amount: number;
  status: PaymentStatus;
  submittedAt: string;
}

export interface PaymentInput {
  userId: number;
  /** @minLength 5 */
  mpesaCode: string;
  /** @minimum 250 */
  amount: number;
}

export interface PlatformStats {
  totalHousekeepers: number;
  totalEmployers: number;
  verifiedHousekeepers: number;
  verifiedEmployers: number;
  countiesCovered: number;
}

export type ListUsersParams = {
/**
 * Filter by role
 */
role?: ListUsersRole;
/**
 * Filter by county/location
 */
county?: string;
minSalary?: number;
maxSalary?: number;
/**
 * Filter by skill
 */
skill?: string;
};

export type ListUsersRole = typeof ListUsersRole[keyof typeof ListUsersRole];


export const ListUsersRole = {
  employer: 'employer',
  housekeeper: 'housekeeper',
} as const;

export type GetFeaturedUsersParams = {
role?: GetFeaturedUsersRole;
};

export type GetFeaturedUsersRole = typeof GetFeaturedUsersRole[keyof typeof GetFeaturedUsersRole];


export const GetFeaturedUsersRole = {
  employer: 'employer',
  housekeeper: 'housekeeper',
} as const;