# lib/api-zod/src/generated/types/payment.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Andy Homecare Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentStatus } from './paymentStatus';

export interface Payment {
  id: number;
  userId: number;
  /** Mpesa transaction code */
  mpesaCode: string;
  /** Amount paid in KES */
  amount: number;
  status: PaymentStatus;
  submittedAt: Date;
}