# lib/api-zod/src/generated/types/paymentInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Andy Homecare Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface PaymentInput {
  userId: number;
  /** @minLength 5 */
  mpesaCode: string;
  /** @minimum 250 */
  amount: number;
}