# lib/api-zod/src/generated/types/platformStats.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Andy Homecare Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface PlatformStats {
  totalHousekeepers: number;
  totalEmployers: number;
  verifiedHousekeepers: number;
  verifiedEmployers: number;
  countiesCovered: number;
}