# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Andy Homecare Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from './userRole';

export interface User {
  id: number;
  fullName: string;
  role: UserRole;
  county: string;
  phone: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  bio?: string | null;
  /** @nullable */
  photoUrl?: string | null;
  /** Monthly salary expectation in KES */
  salaryExpectation: number;
  /**
     * Comma-separated skills (for housekeepers)
     * @nullable
     */
  skills?: string | null;
  /**
     * Years of experience description
     * @nullable
     */
  experience?: string | null;
  /** @nullable */
  languages?: string | null;
  /**
     * Immediate, 2 weeks notice, etc.
     * @nullable
     */
  availability?: string | null;
  paymentVerified: boolean;
  registeredAt: Date;
}