# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Andy Homecare Connect API
 * OpenAPI spec version: 0.1.0
 */
import type { UserInputRole } from './userInputRole';

export interface UserInput {
  /** @minLength 2 */
  fullName: string;
  role: UserInputRole;
  /** @minLength 2 */
  county: string;
  /** @minLength 10 */
  phone: string;
  email?: string;
  bio?: string;
  photoUrl?: string;
  /** @minimum 1 */
  salaryExpectation: number;
  skills?: string;
  experience?: string;
  languages?: string;
  availability?: string;
}