# lib/api-zod/src/generated/types/userUpdate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Andy Homecare Connect API
 * OpenAPI spec version: 0.1.0
 */

export interface UserUpdate {
  /** @minLength 2 */
  fullName?: string;
  county?: string;
  phone?: string;
  email?: string;
  bio?: string;
  photoUrl?: string;
  /** @minimum 1 */
  salaryExpectation?: number;
  skills?: string;
  experience?: string;
  languages?: string;
  availability?: string;
}